BACKPORT: ANDROID: userfaultfd: abort uffdio ops if mmap_lock is contended
Check if the mmap_lock is contended when looping over the pages that
are requested to be filled. When it is observed, we rely on the already
existing mechanism to return bytes copied/filled and -EAGAIN as error.

This helps by avoiding contention of mmap_lock for long running
userfaultfd operations. The userspace can perform other tasks before
retrying the operation for the remaining pages.

Bug: 320478828
Bug: 331535903
Signed-off-by: Lokesh Gidra <lokeshgidra@google.com>
(cherry picked from commit b26824bed88d7ef26dcfc12d6ccc9a5176df6fa1)
[flopster101: Backported to 5.4 userfaultfd code.]
Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>